Refactor(#30-battery-view): BatteryDashboardEnergyView struct로 분리
